+++ sdk/resourcemanager/datafactory/armdatafactory/constants.go
@@ -10,7 +10,7 @@ package armdatafactory
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datafactory/armdatafactory"
-	moduleVersion = "v5.0.0"
+	moduleVersion = "v6.0.0"
 )
 
 // ActivityOnInactiveMarkAs - Status result of the activity when the state is set to Inactive. This is an optional property
@@ -543,6 +543,22 @@ func PossibleGoogleBigQueryAuthenticationTypeValues() []GoogleBigQueryAuthentica
 	return []GoogleBigQueryAuthenticationType{
 		GoogleBigQueryAuthenticationTypeServiceAuthentication,
 		GoogleBigQueryAuthenticationTypeUserAuthentication,
+	}
+}
+
+// GoogleBigQueryV2AuthenticationType - The OAuth 2.0 authentication mechanism used for authentication.
+type GoogleBigQueryV2AuthenticationType string
+
+const (
+	GoogleBigQueryV2AuthenticationTypeServiceAuthentication GoogleBigQueryV2AuthenticationType = "ServiceAuthentication"
+	GoogleBigQueryV2AuthenticationTypeUserAuthentication    GoogleBigQueryV2AuthenticationType = "UserAuthentication"
+)
+
+// PossibleGoogleBigQueryV2AuthenticationTypeValues returns the possible values for the GoogleBigQueryV2AuthenticationType const type.
+func PossibleGoogleBigQueryV2AuthenticationTypeValues() []GoogleBigQueryV2AuthenticationType {
+	return []GoogleBigQueryV2AuthenticationType{
+		GoogleBigQueryV2AuthenticationTypeServiceAuthentication,
+		GoogleBigQueryV2AuthenticationTypeUserAuthentication,
 	}
 }
 
@@ -1515,6 +1531,24 @@ func PossibleSftpAuthenticationTypeValues() []SftpAuthenticationType {
 		SftpAuthenticationTypeBasic,
 		SftpAuthenticationTypeMultiFactor,
 		SftpAuthenticationTypeSSHPublicKey,
+	}
+}
+
+// SnowflakeAuthenticationType - The type used for authentication. Type: string.
+type SnowflakeAuthenticationType string
+
+const (
+	SnowflakeAuthenticationTypeAADServicePrincipal SnowflakeAuthenticationType = "AADServicePrincipal"
+	SnowflakeAuthenticationTypeBasic               SnowflakeAuthenticationType = "Basic"
+	SnowflakeAuthenticationTypeKeyPair             SnowflakeAuthenticationType = "KeyPair"
+)
+
+// PossibleSnowflakeAuthenticationTypeValues returns the possible values for the SnowflakeAuthenticationType const type.
+func PossibleSnowflakeAuthenticationTypeValues() []SnowflakeAuthenticationType {
+	return []SnowflakeAuthenticationType{
+		SnowflakeAuthenticationTypeAADServicePrincipal,
+		SnowflakeAuthenticationTypeBasic,
+		SnowflakeAuthenticationTypeKeyPair,
 	}
 }